Brillouin zones

The first Brillouin zone (FBZ) or full BZ, is the domain of integration for many quantum-mechanical observables in crystalline systems with translation symmetry. In systems with additional point-group symmetries, the symmetries can be used to down-fold the FBZ onto the irreducible Brillouin zone (IBZ) that can be used to reduce the computational cost by a factor of at most $2^d d!$ in $d$ spatial dimensions. While the FBZ is isomorphic to a hypercube, the IBZ is a polytope. The following type stores the data used by various algorithms to do IBZ integration.

AutoBZCore.SymmetricBZ — Type

SymmetricBZ(A, B, lims::AbstractLimits, syms; atol=sqrt(eps()))

Data type representing a Brillouin zone reduced by a set of symmetries, syms with iterated integration limits lims, both of which are assumed to be in the lattice basis (since the Fourier series is). A and B should be identically-sized square matrices containing the real and reciprocal basis vectors in their columns.

Convention

This type assumes all integration limit data is in the reciprocal lattice basis with fractional coordinates, where the FBZ is just the hypercube spanned by the vertices (0,…,0) & (1,…,1). If necessary, use A or B to rotate these quantities into the convention.

lims should be limits compatible with IteratedIntegration.jl. syms should be an iterable collection of point group symmetries compatible with AutoSymPTR.jl.

AutoBZCore.FullBZ — Function

FullBZ(A, B=canonical_reciprocal_basis(A); atol=sqrt(eps()))

Constructs a SymmetricBZ with trivial symmetries.

Integration limits

The IAI algorithm can do IBZ integration with knowledge of the convex hull of the IBZ. The calc_ibz routine in SymmetryReduceBZ.jl does exactly this step. Then polyhedral manipulation algorithms can be used to do this step (search for PolyhedralLimits in IteratedIntegration.jl).

Lattice symmetries

The PTR and AutoPTR algorithms can do IBZ integration (also known as Monkhorst-Pack integration) simply with the crystallographic point group. This can be provided as a vector of the point group operators in the syms argument of SymmetricBZ.

Symmetry representations

The result of the integral on the IBZ still needs to be mapped to the FBZ. The SymRep trait is provided so that the caller can specify the symmetry representation of their integrand to automate this mapping.

AutoBZCore.SymRep — Function

SymRep(f)

SymRep specifies the symmetry representation of the integral of the function f. When you define a new integrand, you can choose to implement this trait to specify how the integral is transformed under the symmetries of the lattice in order to map the integral of f on the IBZ to the result for the FBZ.

New types for SymRep should also extend a corresponding method for AutoBZCore.symmetrize_.

AutoBZCore.AbstractSymRep — Type

AbstractSymRep

Abstract supertype of symmetry representation traits.

AutoBZCore.FaithfulRep — Type

FaithfulRep

Abstract supertype of traits for faithful symmetry representations.

The following representations are provided out-of-the box:

AutoBZCore.TrivialRep — Type

TrivialRep()

Symmetry representation of objects with trivial transformation under the group.

AutoBZCore.LatticeRep — Type

LatticeRep()

Symmetry representation of objects that transform under the group action in the same way as the lattice.

AutoBZCore.UnknownRep — Type

UnknownRep()

Fallback symmetry representation for array types without a user-defined SymRep.

User-defined SymReps

These are the methods to be aware of extending to implement the action of a symmetry group for mapping an integral from the IBZ to the FBZ.

AutoBZCore.symmetrize — Function

symmetrize(f, ::SymmetricBZ, xs...)
symmetrize(f, ::SymmetricBZ, x::Union{Number,AbstractArray{<:Any,0}})

Transform x by the symmetries of the parametrization used to reduce the domain, thus mapping the value of x on the parametrization to the full domain.

AutoBZCore.symmetrize_ — Function

symmetrize_(rep::AbstractSymRep, bz::SymmetricBZ, x)

Transform x under representation rep using the symmetries in bz to obtain the result of an integral on the FBZ from x, which was calculated on the IBZ.

If you are an expert on the symmetry representations of Wannier orbitals or Bloch bands and can offer assistance in automating the IBZ to FBZ mapping, please contact the developers by opening a Github issue or pr.