Density of States

Computing the density of states (DOS) of a self-adjoint, or Hermitian, operator is a related, but distinct problem to the integrals also presented in this package. In fact, many DOS algorithms will compute integrals to approximate the DOS of an operator that depends on a parameter such as crystal momentum. To solve these types of problems, this package defines the following problem type:

AutoBZCore.DOSProblem — Type

DOSProblem(H, domain, [p=NullParameters()])

Define a problem for the density of states of a Hermitian or self-adjoint operator depending on a parameter, H(p), on a given domain in its spectrum. The mathematical definition we use is

\[D(E) = \sum_{k \in p} \sum_{\lambda \in \text{spectrum}(H(k))} \delta(E - \lambda)\]

where $E \in \text{domain}$ and $\delta$ is the Dirac Delta distribution.

Arguments

H: a linear operator depending on a parameter, H(p), that is finite dimensional (e.g: tight binding model) or infinite dimensional (e.g. DFT data)

domain: a set in the spectrum for which an approximation of the density-of-states is desired. Can be a single point, in which case the solution will return the estimated density of states at that eigenvalue, or an interval, in which case the solution will return a function approximation to the density of states on that interval in the spectrum that should be understood as a distribution or measure.

p: optional parameters on which H depends for which the density of states should sum over. Can be discrete (e.g. for H a Hamiltonian with spin degrees of freedom) or continuous (e.g. for H a Hamiltonian parameterized by crystal momentum).

Algorithms

Currently the available algorithms are experimental and we would like to include the following reference algorithms that are more common in the literature in a future release:

(Linear) Tetrahedron Method

Adaptive Gaussian broadening

AutoBZCore.DOSAlgorithm — Type

DOSAlgorithm

Abstract supertype for algorithms for computing density of states

AutoBZCore.GGR — Type

GGR(; npt=50)

Generalized Gilat-Raubenheimer method as in "Generalized Gilat–Raubenheimer method for density-of-states calculation in photonic crystals". This method requires the Hamiltonian and its derivatives, and performs a linear extrapolation at each k-point in an equispace grid. The algorithm is expected to show second-order convergence and suffer reduced error at band crossings compared to interpolatory methods.

Arguments

npt: the number of k-points per dimension

Caching interface

Using the AutoBZCore.init and AutoBZCore.solve! functions, it is possible to construct a cache to solve a DOSProblem for several energies or several Hamiltonians. As an example of solving for several energies,

using AutoBZCore, StaticArrays
h = FourierSeries(SMatrix{1,1,Float64,1}.([0.5, 0.0, 0.5]), period=1.0, offset=-2)
E = 0.3
bz = load_bz(FBZ(), [2pi;;])
prob = DOSProblem(h, E, bz)
alg = GGR(; npt=100)
cache = AutoBZCore.init(prob, alg)
Es = range(-2, 2, length=1000)
data = map(Es) do E
    cache.domain = E
    AutoBZCore.solve!(cache).u
end

As an example of interchanging Hamiltonians, which must remain the same type,

using AutoBZCore, StaticArrays

h = FourierSeries(SMatrix{1,1,Float64,1}.([0.5, 0.0, 0.5]), period=1.0, offset=-2)
bz = load_bz(FBZ(), [2pi;;])
prob = DOSProblem(h, 0.0, bz)
alg = GGR()

cache = AutoBZCore.init(prob, alg)
sol1 = AutoBZCore.solve!(cache)

h.c .*= 2
cache.isfresh = true

sol2 = AutoBZCore.solve!(cache)

sol1.u*2 ≈ sol2.u