Integration

The integration tools provided by AutoBZCore.jl are very flexible and allow for many kinds of user-defined integrals with specialized behaviors. Examples implementations can be found in AutoBZ.jl.

Parametrizing integrals

Integrals.jl provides a generic and unified interface for computing integrals that works well for all kinds of problems. This package builds on that interface by offering IntegralSolvers to easily parametrize an integral for a given integrand, limits of integration, algorithm, and error tolerances. Until the SciML community arrives at a consensus of what a parametrization interface looks like, this gets the job done.

AutoBZCore.IntegralSolver — Type

IntegralSolver(f, lb, ub, alg; abstol=0, reltol=sqrt(eps()), maxiters=typemax(Int))
IntegralSolver(f, bz::SymmetricBZ, alg::AbstractAutoBZAlgorithm; kwargs...)

Constructs a functor that solves an integral of f over the given domain (e.g. lb to ub or a bz) using the given alg to within the given tolerances. Calling this functor, fun with parameters p using the syntax fun(p) returns the estimated integral I. Under the hood, this uses the Integrals.jl interface for defining an IntegralProblem, so f must be a 2-argument function f(x,p), or if in-place, a 3-argument function f(y,x,p).

Also, the types Integrand and FourierIntegrand allow for providing a partial set of parameters so that the IntegralSolver can interface easily with other algorithms, such as root-finding and interpolation.

Integrands

User-defined integrands should be able to take any number of parameters and should be entirely general. We only require that the integrand functions take only positional arguments.

The integrand types below solve the problem of providing the integrand evaluator parameters a partial set of parameters so that the IntegralSolver can parametrize the remaining ones. Namely, the full set of parameters is the union of the ones included in the integrand follow by those passed to the solver, therefore the caller is responsible for ensuring that the ordered union of parameters they pass to the integrand and solver matches the positional arguments of their integrand evaluator. Note that this behavior slightly differs from using the SciML routine remake for an IntegralProblem.

AutoBZCore.Integrand — Type

Integrand(f, p...)

Represent an integrand with a partial collection of parameters p. When the Integrand is invoked with one argument, e.g. int(x), it evaluates f(x, p...). However when invoked with two arguments, as in an IntegralProblem, e.g. int(x, p2), it evaluates the union of parameters f(x, p..., p2...). This allows for convenient parametrization of the integrand.

AutoBZCore.FourierIntegrand — Type

FourierIntegrand(f, s::AbstractFourierSeries, p...)

A type generically representing an integrand f whose entire dependence on the variables of integration is in a Fourier series s, and which may also accept some input parameters p. The caller must know that their function, f, will be evaluated at many points, x, in the following way: f(s(x), p...). Therefore the caller is expected to know the type of s(x) and the layout of the parameters in the tuple p (hint: it should correspond to the arguments of the function). This type is optimized for the IAI and PTR routines.

Algorithms

There are many Brillouin zone integration algorithms, but the ones provided by this package work for black-box integrands and converge automatically to a requested tolerance, except where noted. Additionally, the algorithms are fully compatible with integration over the IBZ, except where noted. More details on the algorithms can be found from their respective packages.

AutoBZCore.AbstractAutoBZAlgorithm — Type

AbstractAutoBZAlgorithm

Abstract supertype for Brillouin zone integration algorithms.

AutoBZCore.IAI — Type

IAI(; order=7, norm=norm, initdivs=nothing, segbufs=nothing)

Iterated-adaptive integration using nested_quadgk from IteratedIntegration.jl. This algorithm is the most efficient for localized integrands. See alloc_segbufs for how to pre-allocate segment buffers for nested_quadgk.

AutoBZCore.PTR — Type

PTR(; npt=50, rule=nothing)

Periodic trapezoidal rule with a fixed number of k-points per dimension, npt, using the routine ptr from AutoSymPTR.jl. The caller should check that the integral is converged w.r.t. npt. See alloc_rule for how to pre-evaluate a PTR rule for use across calls with compatible integrands.

AutoBZCore.AutoPTR — Type

AutoPTR(; norm=norm, buffer=nothing)

Periodic trapezoidal rule with automatic convergence to tolerances passed to the solver with respect to norm using the routine autosymptr from AutoSymPTR.jl. This algorithm is the most efficient for smooth integrands. See alloc_autobuffer for how to pre-evaluate a buffer for autosymptr for use across calls with compatible integrands.

AutoBZCore.PTR_IAI — Type

PTR_IAI(; ptr=PTR(), iai=IAI())

Multi-algorithm that returns an IAI calculation with an abstol determined from the given reltol and a PTR estimate, I, as reltol*norm(I). This addresses the issue that IAI does not currently use a globally-adaptive algorithm and may not have the expected scaling with localization length unless an abstol is used since computational effort may be wasted via a reltol with the naive nested_quadgk.

AutoBZCore.AutoPTR_IAI — Type

AutoPTR_IAI(; reltol=1.0, ptr=AutoPTR(), iai=IAI())

Multi-algorithm that returns an IAI calculation with an abstol determined from an AutoPTR estimate, I, computed to reltol precision, and the rtol given to the solver as rtol*norm(I). This addresses the issue that IAI does not currently use a globally-adaptive algorithm and may not have the expected scaling with localization length unless an abstol is used since computational effort may be wasted via a reltol with the naive nested_quadgk.

AutoBZCore.TAI — Type

TAI(; rule=HCubatureJL())

Tree-adaptive integration using hcubature from HCubature.jl. This routine is limited to integration over hypercube domains and may not use all symmetries.

Allocations for algorithms

In practice, allocating memory for an algorithm can help improve its performance. This is particularly noticeable for PTR, where one of the main optimizations is to store the Fourier series evaluated on the PTR grid for reuse across multiple compatible calls. The routines below show how to pre-evaluate or pre-allocate the memory necessary for different algorithms.

AutoBZCore.alloc_rule — Function

alloc_rule(f::AbstractFourierSeries, ::Type, syms)

Compute the values of f on the PTR grid as well as the quadrature weights for the given syms to use across multiple compatible calls of the PTR algorithm.

AutoSymPTR.alloc_autobuffer — Function

alloc_autobuffer(f::AbstractFourierSeries, ::Type, syms)

Initialize an empty buffer of PTR rules with pre-evaluated Fourier series evaluated on a domain of type T with symmetries syms to use across multiple compatible calls of the AutoPTR algorithm.

IteratedIntegration.alloc_segbufs — Function

alloc_segbufs(coefficient_type, range_type, norm_type, ndim)

Allocate a segment buffer for multiple compatible calls to IAI. Typically coefficient_type would be eltype(bz), range_type would be the return type of the integrand, and norm_type would be Base.promote_op(norm, range_type). ndim should be the number of dimensions of integration.

Parallelization

This package parallelizes $k$-sums in the PTR and AutoPTR routines by default, which gives good multi-threaded performance for those algorithms when the number of $k$-points is large. On the other hand, if the number of parameter points for which you would like to evaluate an integral is large, then parameter-parallelization is an effective strategy and done by the routines below.

AutoBZCore.batchsolve — Function

batchsolve(f, ps, [T=Base.promote_op(f, eltype(ps))]; nthreads=Threads.nthreads())

Evaluate the IntegralSolver f at each of the parameters ps in parallel. Returns a vector containing the evaluated integrals I. This is a form of multithreaded broadcasting.

AutoBZCore.batchparam — Function

batchparam(ps, nthreads)

If the cost of a calculation smoothly varies with the parameters ps, then batch ps into nthreads groups where the ith element of group j is ps[j+(i-1)*nthreads]