Reference

AutoBZCore.FourierIntegrand — Type

FourierIntegrand(f, s::AbstractFourierSeries, p...)

A type generically representing an integrand f whose entire dependence on the variables of integration is in a Fourier series s, and which may also accept some input parameters p. The caller must know that their function, f, will be evaluated at many points, x, in the following way: f(s(x), p...). Therefore the caller is expected to know the type of s(x) (hint: eltype(s)) and the layout of the parameters in the tuple p (hint: it should correspond to the arguments of the function). This type is optimized for the IAI and PTR routines.

AutoBZCore.SymmetricBZ — Type

SymmetricBZ(A, B, lims::AbstractLimits, syms; atol=sqrt(eps()))

Data type representing a Brillouin zone reduced by a set of symmetries, syms with iterated integration limits lims, both of which are assumed to be in the lattice basis (since the Fourier series is). A and B should be identically-sized square matrices containing the real and reciprocal basis vectors in their columns. lims should be limits compatible with the IteratedIntegration package that represent the BZ in fractional lattice coordinates (e.g. the full BZ with vertices (0,0,0) & (1,1,1)). syms should be a collection of symmetries compatible with AutoSymPTR and the symmetry operators should be in the lattice basis (if necessary, rotate them from the Cartesian basis).

AutoBZCore.FullBZ — Function

FullBZ(A, B=canonical_reciprocal_basis(A), lims=lattice_bz_limits(B); atol=sqrt(eps()))

A type alias for SymmetricBZ{Nothing} when there are no symmetries applied to BZ

AutoBZCore.batchparam — Method

batchparam(xs, nthreads)

If the cost of a calculation smoothly varies with the parameters xs, then batch xs into nthreads groups where the ith element of group j is xs[j+(i-1)*nthreads]

AutoBZCore.batchsolve — Function

batchsolve(f, ps, [T=Base.promote_op(f, eltype(ps))]; nthreads=Threads.nthreads())

Evaluate the IntegralSolver f at each of the parameters ps in parallel. Returns a vector containing the evaluated integrals I. This is a form of multithreaded broadcasting.

AutoBZCore.symmetrize — Method

symmetrize(f, ::SymmetricBZ, xs...)
symmetrize(f, ::SymmetricBZ, x::Number)

Tranform x by the symmetries of the parametrization used to reduce the domain, thus mapping the value of x on the parametrization to the full domain.