AutoBZCore.jl

AutoBZCore — Module

The package providing the functionality and abstractions for AutoBZ.jl. It provides both SymmetricBZ, which is a type that stores information about the Brillouin zone (BZ), and IntegralSolver, a type that provides a functor interface to parametrize IntegralProblems as defined by Integrals.jl. The package also provides the optimized FourierIntegrand. The package integrates with FourierSeriesEvaluators.jl, IteratedIntegration.jl, and AutoSymPTR.jl to provide a generic interface to efficient algorithms for BZ integration.

For example, computing the local Green's function can be done as follows:

using LinearAlgebra
using FourierSeriesEvaluators
using AutoBZCore

gloc_integrand(h_k, η, ω) = inv(complex(ω,η)*I-h_k)     # define integrand evaluator
h = FourierSeries([0.5, 0.0, 0.5]; offset=-2)           # construct cos(k) 1D integer lattice Hamiltonian
bz = FullBZ(2pi*I(1))                                   # construct BZ from lattice vectors A=2pi*I
integrand = FourierIntegrand(gloc_integrand, h, 0.1)    # construct integrand with Fourier series h and parameter η=0.1
alg = IAI()                                             # choose integration algorithm (also AutoPTR() and PTR())
gloc = IntegralSolver(integrand, bz, alg; abstol=1e-3)  # construct a solver for gloc to within specified tolerance
gloc(0.0)                                               # evaluate gloc at frequency ω=0.0

Assumptions

AutoBZCore assumes that all calculations occur in the reciprocal lattice basis, since that is the basis in which Wannier interpolants are most efficiently described. See SymmetricBZ for details.