AutoBZCore.jl

AutoBZCore — Module

The package providing the functionality and abstractions for AutoBZ.jl. It provides both SymmetricBZ, which is a type that stores information about the Brillouin zone (BZ), and FourierIntegrator, a type that provides a functor interface to compute user-defined Brillouin zone integrals depending on Wannier-interpolated quantities. The package integrates with FourierSeriesEvaluators, IteratedIntegration, and AutoSymPTR to provide a generic interface to efficient algorithms for BZ integration.

For example, computing the local Green's function can be done as follows:

using LinearAlgebra
using FourierSeriesEvaluators
using AutoBZCore

gloc_integrand(H, η, ω) = inv(complex(ω,η)*I-H) # define integrand evaluator
h = FourierSeries([0.5, 0.0, 0.5]; offset=-2) # define 1D integer lattice Hamiltonian
fbz = FullBZ(I(1)) # construct BZ from lattice with default limits of integration
ps = (1.0, 0.0) # representative values for (η,ω), the last arguments of the integrand evaluator
gloc = FourierIntegrator(gloc_integrand, fbz, h; ps=ps) # initialize default integration routine
gloc(ps...) # evaluate gloc at parameter points

Assumptions

AutoBZCore assumes that all calculations occur in the reciprocal lattice basis, since that is the basis in which Wannier interpolants are most efficiently described. This means that user-provided integration limits and symmetries should be in the reciprocal lattice basis and additionally they should be in fractional coordinates (e.g. the BZ in these coordinates has vertices (0,0,0) and (1,1,1)). All symmetry transformations must be dealt with by the user, who can specialize the symmetrize routine to automate that step.