Function reference

The following are functions exported by AutoBZCore.jl

AutoBZCore.load_bz — Function

load_bz(::AbstractBZ, T::Type)
load_bz(::AbstractBZ, A::AbstractMatrix, [B::AbstractMatrix])

Interface to loading Brillouin zones

Assumptions

AutoBZCore assumes that all calculations occur in the reciprocal lattice basis, since that is the basis in which Wannier interpolants are most efficiently described. See SymmetricBZ for details. We also assume that the integrands are cheap to evaluate, which is why we provide adaptive methods in the first place, so that return types can be determined at runtime (and mechanisms are in place for compile time as well)

AutoBZCore.batchsolve — Function

batchsolve(f::IntegralSolver, ps::AbstractArray, [T]; nthreads=Threads.nthreads())

Evaluate the IntegralSolver f at each of the parameters ps in parallel. Returns an array similar to ps containing the evaluated integrals I. This is a form of multithreaded broadcasting. Providing the return type f(eltype(ps))::T is optional, but will help in case inference of that type fails.

Missing docstring.

Missing docstring for AutoBZCore.paramzip. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZCore.paramproduct. Check Documenter's build log for details.

Brillouin-zone kinds

AutoBZCore.FBZ — Type

FBZ{N} <: AbstractBZ

Singleton type representing first/full Brillouin zones of N dimensions. By default, N is nothing and the dimension is obtained from input files.

AutoBZCore.IBZ — Type

IBZ <: AbstractBZ

Singleton type representing irreducible Brillouin zones. Load SymmetryReduceBZ.jl to use this.

AutoBZCore.InversionSymIBZ — Type

InversionSymIBZ{N} <: AbstractBZ

Singleton type representing Brillouin zones with full inversion symmetry

Assumptions

Only expect this to work for systems with orthogonal lattice vectors

AutoBZCore.CubicSymIBZ — Type

CubicSymIBZ{N} <: AbstractBZ

Singleton type representing Brillouin zones with full cubic symmetry

Assumptions

Only expect this to work for systems with orthogonal lattice vectors

Symmetry representations

AutoBZCore.SymRep — Function

SymRep(f)

SymRep specifies the symmetry representation of the integral of the function f. When you define a new integrand, you can choose to implement this trait to specify how the integral is transformed under the symmetries of the lattice in order to map the integral of f on the IBZ to the result for the FBZ.

New types for SymRep should also extend a corresponding method for AutoBZCore.symmetrize_.

AutoBZCore.TrivialRep — Type

TrivialRep()

Symmetry representation of objects with trivial transformation under the group.

AutoBZCore.UnknownRep — Type

UnknownRep()

Fallback symmetry representation for array types without a user-defined SymRep.